% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_table.R
\name{descriptive_table}
\alias{descriptive_table}
\title{Mean, SD, and Correlation table}
\usage{
descriptive_table(
  data,
  cols,
  cor_sig_test = "raw",
  cor_digit = 3,
  mean_sd_digit = 3,
  filepath = NULL
)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) need to be included in the table.}

\item{cor_sig_test}{adjusted or raw. Default as adjusted. See psych::corr.test to learn more.}

\item{cor_digit}{number of digit for correlation table}

\item{mean_sd_digit}{number of digit for mean and sd tables}

\item{filepath}{provide full path to pass into the write.csv(file = filepath)}
}
\value{

}
\description{
This function uses the psych::corr.test function to generated the correlation. One potential problem with the current version of the function is that if the correlation between the two item is too high (rounded as 1), then that cell will become blank. This is a very rare problem, and I have not yet figure out how to solve this.
}
\examples{
descriptive_table(iris,cols = tidyr::everything()) # all columns
descriptive_table(iris,cols = where(is.numeric)) # all numeric columns
}
